Some buildsys changes from Johannes Kellner, many thanks to him. Here's what he has to say about them:
New Features:

Targets have now the new command: 'changedllextension ext'
So you can change the set the extension of a dll to somthing else than
dll. If not set the default will remain 'dll'. This function is only
evaluated for dll targettypes.
This is needed to create dll's with special extensions, as required for
plugins for max or maya.

Modules have now a new command: 'setmoddeffile filename'
So you can setup Module Defintion File, whats a other way to define
dllexports. This is needed for the max plugins, because max want's some
dllexports that must be published this way.
The file must be in src/$moduledir. It's not valid to have more than one
module with a module def file set in a target. (Only the first found
will be used, but it's possible to have more than one for a target.)

Changes:
I have added output for the new functions to the dump_date and
dump_api_data procs.

I have changed nearly all functions in loadbld.tcl:
I have replaced the constant numbers of the diffent block types with
variables. It's mutch more readable now I mean.
Also I have made all functions save, so they will report a failure if a
function is used inside a wrong block type. (for example if 'settarget
foo' in a bundle block) Such a error is now reported and the function
don't changes anything.
Also I changed the 'end....' commands, so they will now hard fail if the
wrong one is used. I thik this must be because if a false 'end...' is
used everything gets invalid.

Open Tasks:
I can't do the needed things for the vc6 generator, can please someone
add the new features.
I changed the vc71 generator and expect that the changes will work, but
I don't have the option to test this, so can please someone valid it.