nParticleShapeNode's can now be moved around.
Moved the emitter setup code from Attach() to RenderTransform(), this was necessary to allow the node's matrix to be updated by animators and/or script commands. This should be ok but hasn't been tested with multiple instances yet, I might have to whip up a test shell for attaching nodes multiple times.