A simple example of nnetwork usage, from Luis Cabellos (Tragnarion Studios).
Note: I changed this slightly to be part of the nnetwork contrib instead of a seperate contribution. It has some trouble with the project files, and crashes at startup. Those problems may well be something to do with the changes I made but I wanted to get it in so that others could sort it out if it takes me too long.